refactor: streamline expense dialog component and update related imports
