+++ app/components/vehicle/expense/VehicleExpenseFuelStatCard.vue
@@ -108,7 +108,6 @@ const periodOptions = [
           :ui="{
             list: 'ring ring-accented rounded-full',
             indicator: 'rounded-full',
-            trigger: 'w-1/2',
           }"
         />
       </div>
