+++ app/components/vehicle/expense/VehicleExpenseStatCard.vue
@@ -154,7 +154,6 @@ const uniqueUnits = computed(() => {
             :ui="{
               list: 'ring ring-accented rounded-full',
               indicator: 'rounded-full',
-              trigger: 'w-1/2',
             }"
           />
