+++ app/pages/vehicles/[id]/index.vue
@@ -1,9 +1,11 @@
 <script setup lang="ts">
 definePageMeta({
   layout: "vehicle",
-  redirect: (route) => `/vehicles/${route.params.id}/expenses`,
+  //redirect: (route) => `/vehicles/${route.params.id}/expenses`,
 });
 </script>
 <template>
-  <NuxtPage />
+  <UPage class="flex-1 min-h-0">
+    <VehicleCard />
+  </UPage>
 </template>
