+++ app/pages/vehicles/[id]/stats/index.vue
@@ -4,7 +4,6 @@ useHead({
 });
 definePageMeta({
   auth: true,
-  layout: "vehicle",
 });
 </script>
